requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
djangorestframework==3.16.1
filelock==3.19.1
fsspec==2025.7.0
gunicorn==23.0.0
h11==0.16.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.2
numpy==2.0.2
packaging==25.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uuid==1.30
uvicorn==0.35.0
whitenoise==6.9.0
